+++ sdk/storage/azure-storage-queue/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.2.1 (Unreleased)
+## 12.3.0b1 (Unreleased)
 
 ### Features Added
 - Added support for `create_if_not_exists()` for `QueueClient`